pyproject.toml
[tool.poetry]
name = "Minute Inbox"
version = "0.1.2"
description = "Get a disposable 10 minute email address"
documentation = "https://github.com/fedecalendino/alfred-minute-inbox/blob/main/README.md"
homepage = "https://github.com/fedecalendino/alfred-minute-inbox"
license = "MIT"

authors = [
    "Fede Calendino <fede@calendino.com>"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
alfred-pyflow = "0.3.3"
minuteinbox = "0.3.0"
beautifulsoup4 = "^4.11.1"
black = "^22.12.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
